# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AllYourBase API
  description: |
    Backend-as-a-Service for PostgreSQL. Single binary. Auto-generated REST API
    with filtering, sorting, pagination, foreign key expansion, full-text search,
    batch operations, authentication, file storage, realtime events, and webhooks.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: AllYourBase
    url: https://allyourbase.io
servers:
  - url: http://localhost:8090
    description: Local development server

# Default: no auth required. Individual operations override with BearerAuth or AdminAuth.
security:
  - {}
...
paths:
...
  /api/collections/{table}:
    get:
      tags: [Collections]
      summary: List records
      operationId: listRecords
      parameters:
        - $ref: "#/components/parameters/TablePath"
        - $ref: "#/components/parameters/Filter"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Fields"
        - $ref: "#/components/parameters/Expand"
        - $ref: "#/components/parameters/SkipTotal"
        - $ref: "#/components/parameters/Search"
      security:
        - BearerAuth: []
        - {}
      responses:
        "200":
          description: Paginated list of records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListResponse"
        "400":
          description: Invalid filter or search syntax
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Collections]
      summary: Create a record
      operationId: createRecord
      parameters:
        - $ref: "#/components/parameters/TablePath"
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Record fields as key-value pairs
              additionalProperties: true
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                type: object
                description: The created record
                additionalProperties: true
        "400":
          description: Invalid JSON or no recognized columns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Unique constraint violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: NOT NULL or check constraint violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/collections/{table}/batch:
    post:
      tags: [Collections]
      summary: Batch operations
      description: |
        Perform multiple create, update, and delete operations in a single
        atomic transaction. If any operation fails, all changes are rolled back.
        Max 1000 operations per request.
      operationId: batchOperations
      parameters:
        - $ref: "#/components/parameters/TablePath"
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequest"
      responses:
        "200":
          description: Batch results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BatchResult"
        "400":
          description: Invalid request or operation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/collections/{table}/{id}:
    get:
      tags: [Collections]
      summary: Get a record
      operationId: getRecord
      parameters:
        - $ref: "#/components/parameters/TablePath"
        - $ref: "#/components/parameters/RecordId"
        - $ref: "#/components/parameters/Fields"
        - $ref: "#/components/parameters/Expand"
      security:
        - BearerAuth: []
        - {}
      responses:
        "200":
          description: The record
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Collection or record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Collections]
      summary: Update a record (partial)
      operationId: updateRecord
      parameters:
        - $ref: "#/components/parameters/TablePath"
        - $ref: "#/components/parameters/RecordId"
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Fields to update
              additionalProperties: true
      responses:
        "200":
          description: The updated record
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid JSON or no recognized columns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Collection or record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Collections]
      summary: Delete a record
      operationId: deleteRecord
      parameters:
        - $ref: "#/components/parameters/TablePath"
        - $ref: "#/components/parameters/RecordId"
      security:
        - BearerAuth: []
        - {}
      responses:
        "204":
          description: Record deleted
        "404":
          description: Collection or record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/rpc/{function}:
    post:
      tags: [RPC]
      summary: Call a PostgreSQL function
      description: |
        Invoke a PostgreSQL function by name with named arguments. Void functions
        return 204 No Content. Set-returning functions return an array. Scalar
        functions return the unwrapped value.
      operationId: callFunction
      parameters:
        - name: function
          in: path
          required: true
          description: PostgreSQL function name
          schema:
            type: string
      security:
        - BearerAuth: []
        - {}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Named arguments as key-value pairs
              additionalProperties: true
      responses:
        "200":
          description: Function result (scalar value, object, or array)
          content:
            application/json:
              schema: {}
        "204":
          description: Void function executed successfully
        "400":
          description: Invalid arguments or function error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Function not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/login or /api/auth/register
    AdminAuth:
      type: http
      scheme: bearer
      description: Admin token from /api/admin/auth
  parameters:
    TablePath:
      name: table
      in: path
      required: true
      description: Table/collection name
      schema:
        type: string
    RecordId:
      name: id
      in: path
      required: true
      description: Record primary key value (composite keys separated by commas)
      schema:
        type: string
...
    Filter:
      name: filter
      in: query
      description: "SQL-safe filter expression (e.g. status='active' AND age>21)"
      schema:
        type: string
    Sort:
      name: sort
      in: query
      description: "Sort fields with direction prefix (e.g. -created_at,+title). - = DESC, + = ASC"
      schema:
        type: string
    Page:
      name: page
      in: query
      description: Page number (default 1)
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: perPage
      in: query
      description: Items per page (default 20, max 500)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 500
    Fields:
      name: fields
      in: query
      description: "Comma-separated column names to return (e.g. id,name,email)"
      schema:
        type: string
    Expand:
      name: expand
      in: query
      description: "Comma-separated foreign key column names to expand (e.g. author,category)"
      schema:
        type: string
    SkipTotal:
      name: skipTotal
      in: query
      description: Skip COUNT query for faster list responses
      schema:
        type: boolean
        default: false
    Search:
      name: search
      in: query
      description: "Full-text search across all text columns. Supports PostgreSQL websearch_to_tsquery syntax."
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "collection not found: nonexistent"
        data:
          type: object
          description: Field-level validation details
          additionalProperties: true
    ListResponse:
      type: object
      required: [items, page, perPage, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
          description: Records for the current page
        page:
          type: integer
          example: 1
        perPage:
          type: integer
          example: 20
        totalItems:
          type: integer
          description: Total matching records (-1 if skipTotal=true)
          example: 42
        totalPages:
          type: integer
          description: Total pages (-1 if skipTotal=true)
          example: 3
...
    BatchRequest:
      type: object
      required: [operations]
      properties:
        operations:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/BatchOperation"
    BatchOperation:
      type: object
      required: [method]
      properties:
        method:
          type: string
          enum: [create, update, delete]
        id:
          type: string
          description: Record ID (required for update and delete)
        body:
          type: object
          description: Record data (required for create and update)
          additionalProperties: true
    BatchResult:
      type: object
      required: [index, status]
      properties:
        index:
          type: integer
          description: Position of the operation in the request array
        status:
          type: integer
          description: HTTP status code for this operation (201, 200, or 204)
        body:
          type: object
          description: The created or updated record (absent for deletes)
          additionalProperties: true
...